Move private validation types to separate module
